requirements.txt
gradio
google-cloud-aiplatform
httpx
